When evaluating an 'H' Block for neighbour resolution:
Now treat a 'C' Block immediately before/after, as if that 'H' Block were
at the beginning/end (or both) of the String.

This is because a 'C' Block can not meaningfully participate in the 'are
we surrounded by A/B Blocks ?' decisions, so needs to be treated as
'no block'.

Previously, this part of code was leaving such 'H' Blocks unresolved, to
fall to a default resolution later on, which was causing inefficient
encodings due to unnecessary code changes.

In practice, it isn't just 'C' Blocks; it's all Blocks other than A/B